Changed the alarm system so that the user can login and report his/her attendance
Completely got rid of the banking system. It is more of a Supervisor - Student concept now. Even when the user closes the app, the alarm will still be recorded. Added in some inspirational quotes for users when they turn off the alarm. The history page records all past records of the user. Including when he/she turned off the alarm without penalty.